Modified fields on almuerzo.almuerzo model
The cedula field string now has an accent mark.
The employee_object field was removed because it is not necessary on this model and the access to eat of the employee is validated on almuerzo.despacho.almuerzo model
The unit field now has a lambda function that set the department Dirección Administrativa as is requested on bussiness rules.
The comment field was added.