Bug fixed after refactor code for almuerzo.almuerzo and almuerzo.despacho.almuerzo
almuerzo.py
unit field is now a many2one field.
rec.unit now is set to bring the id instead of the name because unit was changed to a many2one field.

despacho_almuerzo.py
refactor the call to the function _compute_rec_name() to _compute_full_name

almuerzo_views.xml
unit field was change to not allow creation, edit and open.

despacho_almuerzo.xml
field name and unit now include the attribute force_save to force the saving of those attributes on the model.